Add job-related interfaces and logging service for scraper functionality
